backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.51.0
anyio==4.9.0
argcomplete==3.6.2
asyncio==3.4.3
attrs==25.3.0
beautifulsoup4==4.13.4
boto3==1.38.14
botocore==1.38.14
bs4==0.0.2
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
cohere==5.15.0
colorama==0.4.6
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
eval_type_backport==0.2.2
executing==2.2.0
fastapi==0.115.12
fastavro==1.10.0
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.3.2
google-auth==2.40.1
googleapis-common-protos==1.70.0
gotrue==2.12.0
griffe==1.7.3
groq==0.24.0
h11==0.16.0
h2==4.2.0
hf-xet==1.1.1
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.31.1
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.1.0
jiter==0.9.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-core==0.3.59
langchain-text-splitters==0.3.8
langsmith==0.3.42
logfire==3.15.1
logfire-api==3.15.1
markdown-it-py==3.0.0
mcp==1.8.1
mdurl==0.1.2
mistralai==1.7.0
multidict==6.4.3
openai==1.78.1
opentelemetry-api==1.33.0
opentelemetry-exporter-otlp-proto-common==1.33.0
opentelemetry-exporter-otlp-proto-http==1.33.0
opentelemetry-instrumentation==0.54b0
opentelemetry-proto==1.33.0
opentelemetry-sdk==1.33.0
opentelemetry-semantic-conventions==0.54b0
orjson==3.10.18
packaging==24.2
pluggy==1.5.0
postgrest==1.0.1
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic-ai==0.2.0
pydantic-ai-slim==0.2.0
pydantic-evals==0.2.0
pydantic-graph==0.2.0
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
realtime==2.4.3
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rsa==4.9.1
s3transfer==0.12.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
sse-starlette==2.3.5
starlette==0.46.2
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.1
supafunc==0.9.4
tenacity==9.1.2
tokenizers==0.21.1
tqdm==4.67.1
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
wcwidth==0.2.13
websockets==14.2
wrapt==1.17.2
yarl==1.20.0
zipp==3.21.0
zstandard==0.23.0
